fix(sqlservice): Upgrade SQLService to allow use of pre-ping
Due to connection pooling issues with gunicorn, at times the database
connections time out, which causes an 'MySQL server has gone away'
errors for the clients.

Firstly, we set the connection pool recycle time, but more importantly
we are also enabling the pre-ping functionality. This causes SQLAlchemy
to test each connection are they are being activated from the pool, and
in the cause of an error, it will attempt to reconnect to the DB server
before returning the connection