feat(templates): Move templates to database
In order to allow users to edit the templates being used for
notifications etc, instead of loading templates directly from disk
instead we no load the templates from disk into the database and when a
subsystem needs a template they will pull it from the database instead
of from disk.

On every startup of the system, a check will be performed comparing
templates on disk vs the templates in database. If any new templates or
any changes are detected, the template will be re-imported. This however
will only happen provided the user has NOT manually modified the
template through the admin interface. Any user edited templates will not
be automatically overwritten, but there is a function in the admin
console to allow users to overwrite templates