refactor: Refactored Account objects
Updated Account objects to be a generic type instead of using the DB
schema directly. This allows for much easier adaptation to other non-AWS
cloud providers.

This change also modifies the DNS collector to use a
new AXFR or CloudFlare account type object with the settings on the
account object. This allows you to configure more than one AXFR and/or
CloudFlare accounts